Rename service to application, fix genesis block parameter passing
