+++ irohad/main/CMakeLists.txt
@@ -31,7 +31,7 @@ target_link_libraries(raw_block_insertion
     )
 
 add_library(application
-    service.cpp
+        application.cpp
     impl/ordering_init.cpp
     impl/consensus_init.cpp
     impl/block_loader_init.cpp
